Restructure files to be more appropriate (#1131)
Mostly just moving files and functions around to be more appropriate.
