+++ nativelink-scheduler/src/memory_scheduler_state/awaited_action_db/sorted_awaited_action.rs
@@ -1,0 +1,46 @@
+// Copyright 2024 The NativeLink Authors. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::cmp;
+use std::sync::Arc;
+
+use super::awaited_action::{AwaitedAction, AwaitedActionSortKey};
+
+#[derive(Debug, Clone)]
+pub struct SortedAwaitedAction {
+    pub sort_key: AwaitedActionSortKey,
+    pub awaited_action: Arc<AwaitedAction>,
+}
+
+impl PartialEq for SortedAwaitedAction {
+    fn eq(&self, other: &Self) -> bool {
+        Arc::ptr_eq(&self.awaited_action, &other.awaited_action) && self.sort_key == other.sort_key
+    }
+}
+
+impl Eq for SortedAwaitedAction {}
+
+impl PartialOrd for SortedAwaitedAction {
+    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for SortedAwaitedAction {
+    fn cmp(&self, other: &Self) -> cmp::Ordering {
+        self.sort_key.cmp(&other.sort_key).then_with(|| {
+            Arc::as_ptr(&self.awaited_action).cmp(&Arc::as_ptr(&other.awaited_action))
+        })
+    }
+}
