+++ nativelink-util/Cargo.toml
@@ -10,6 +10,7 @@ nativelink-proto = { path = "../nativelink-proto" }
 
 async-lock = "3.3.0"
 async-trait = "0.1.80"
+bitflags = "2.5.0"
 blake3 = { version = "1.5.1", features = ["mmap"] }
 bytes = "1.6.0"
 console-subscriber = { version = "0.3.0" }
