+++ nativelink-util/src/lib.rs
@@ -23,6 +23,7 @@ pub mod fastcdc;
 pub mod fs;
 pub mod health_utils;
 pub mod metrics_utils;
+pub mod operation_state_manager;
 pub mod origin_context;
 pub mod platform_properties;
 pub mod proto_stream_utils;
